# did-server/src/main/java/com/managination/numa/didserver/dto/VersionConflictResponse.java
package com.managination.numa.didserver.dto;

import com.managination.numa.didserver.model.DidDocument;

public record VersionConflictResponse(
    String error,
    String message,
    String serverVersionId,
    String clientVersionId,
    DidDocument serverDocument
) {}