# did-server/src/main/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DID Server API",
    "description": "API for managing Decentralized Identifiers (DIDs) using the did:webvh method, credential issuance sessions, verifiable presentations, and reward claiming.\n\n## DID Format\n\nThis server supports the `did:webvh` method with SCID (Self-Contained Identifier).\n\n**Format:** `did:webvh:SCID:domain[:path...]`\n\n**Example:** `did:webvh:SCID:example.com:.well-known`\n\n## Storage Structure\n\nDID files are stored as `did.jsonl` in the following structure:\n- Root domain: `{storagePath}/{domain}/.well-known/did.jsonl`\n- Sub-path: `{storagePath}/{domain}/{path}/did.jsonl`\n\n## WebSocket Sessions\n\nCredential issuance uses WebSocket sessions for real-time communication between issuer and holder. See `did-server-asyncapi.json` for the complete WebSocket channel specification.",
    "version": "2.0.0",
    "contact": {
      "name": "Swiss Federal Chancellery"
    },
    "license": {
      "name": "Apache License 2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://did.ninja",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/dids": {
      "post": {
        "tags": ["DID Management"],
        "summary": "Upload a DID log file",
        "description": "Upload and verify a DID log file in JSONL format. The server performs the following validations:\n\n1. Ensures the uploaded file is not empty\n2. Extracts the DID ID and verifies the domain matches the `Host` header\n3. If a file already exists at the target path, verifies the SCID matches\n4. Resolves the DID log to verify its integrity\n\nOn successful validation, the file is stored at the resolved path.",
        "operationId": "uploadDid",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "didJsonl": {
                    "type": "string",
                    "format": "binary",
                    "description": "The DID log file in JSONL format."
                  }
                },
                "required": ["didJsonl"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DID uploaded and verified successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "success": {
                    "value": "DID uploaded and verified successfully: did:webvh:SCID:example.com"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/did": {
      "post": {
        "tags": ["DID Management"],
        "summary": "Register a new DID with the server",
        "description": "Register a newly created did:webvh identity with the server. The client sends the DID string, DID document, and the complete DID log. The server validates the DID format, verifies the document integrity, and stores it.\n\nThis endpoint is used by `DidManager.createDid()` after local DID creation to persist the identity on the server.\n\n### Validation Steps\n\n1. Validates DID format (must be `did:webvh:...`)\n2. Verifies the DID document structure\n3. Validates the DID log entries and proof chain\n4. Stores the DID document and log",
        "operationId": "registerDid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DidRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DID registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid DID registration request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DID already exists with different document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/did/{did}": {
      "get": {
        "tags": ["DID Management"],
        "summary": "Resolve a DID to its document",
        "description": "Resolve a did:webvh identifier to its current DID document. The server looks up the DID in storage and returns the latest version of the DID document.\n\nThis endpoint is used by `DidManager.resolveDid()` to fetch remote DID documents.",
        "operationId": "resolveDid",
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "description": "URL-encoded DID string (e.g., `did%3Awebvh%3ASCID%3Aexample.com`)",
            "schema": {
              "type": "string",
              "pattern": "^did:webvh:.+"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DID document found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidDocument"
                }
              }
            }
          },
          "404": {
            "description": "DID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/did/{did}/update": {
      "put": {
        "tags": ["DID Management"],
        "summary": "Update an existing DID document",
        "description": "Update an existing DID document with a new version. The client must send the updated DID document along with the new log entry. The server validates that the update is authorized by checking the proof chain and that the client possesses the appropriate signing key.\n\n### Sync Requirement\n\nBefore calling this endpoint, the client should call `GET /did/{did}` to fetch the current version and ensure the local copy matches. If the server version differs, the client must merge changes before submitting the update.",
        "operationId": "updateDid",
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "description": "URL-encoded DID string",
            "schema": {
              "type": "string",
              "pattern": "^did:webvh:.+"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DidUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DID document updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized - proof verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict - server has newer version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionConflictResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "OAuth 2.0 Bearer token obtained from the /token endpoint"
      }
    },
    "schemas": {
...
      "DidDocument": {
        "type": "object",
        "description": "W3C DID Document",
        "properties": {
          "@context": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "id": {
            "type": "string",
            "description": "DID identifier (e.g., did:webvh:SCID:example.com)"
          },
          "controller": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ],
            "description": "Entity that controls the DID. Includes the server's public key for recovery."
          },
          "alsoKnownAs": {
            "type": "array",
            "items": { "type": "string" }
          },
          "verificationMethod": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerificationMethod"
            }
          },
          "authentication": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string" },
                { "$ref": "#/components/schemas/VerificationMethod" }
              ]
            }
          },
          "assertionMethod": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string" },
                { "$ref": "#/components/schemas/VerificationMethod" }
              ]
            }
          },
          "keyAgreement": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string" },
                { "$ref": "#/components/schemas/VerificationMethod" }
              ]
            }
          },
          "capabilityInvocation": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string" },
                { "$ref": "#/components/schemas/VerificationMethod" }
              ]
            }
          },
          "capabilityDelegation": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string" },
                { "$ref": "#/components/schemas/VerificationMethod" }
              ]
            }
          },
          "service": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceEndpoint"
            }
          }
        },
        "required": ["id"]
      },
      "VerificationMethod": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "controller": { "type": "string" },
          "publicKeyMultibase": { "type": "string" },
          "publicKeyJwk": {
            "type": "object",
            "properties": {
              "kty": { "type": "string" },
              "crv": { "type": "string" },
              "x": { "type": "string" },
              "y": { "type": "string" }
            }
          }
        },
        "required": ["id", "type", "controller"]
      },
      "ServiceEndpoint": {
        "type": "object",
        "description": "Service endpoint in a DID document. Used to store the EVM address for NFT issuance.",
        "properties": {
          "id": { "type": "string" },
          "type": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "serviceEndpoint": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } },
              { "type": "object" }
            ],
            "description": "For EVM addresses: { \"evmAddress\": \"0x...\" }"
          }
        },
        "required": ["id", "type", "serviceEndpoint"]
      },
      "DidRegistrationRequest": {
        "type": "object",
        "description": "Request to register a new DID with the server.",
        "properties": {
          "did": {
            "type": "string",
            "description": "The DID string (e.g., did:webvh:SCID:example.com)"
          },
          "document": {
            "$ref": "#/components/schemas/DidDocument",
            "description": "The DID document"
          },
          "log": {
            "type": "array",
            "description": "DID log entries with proofs",
            "items": {
              "type": "object"
            }
          }
        },
        "required": ["did", "document", "log"]
      },
      "DidRegistrationResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "did": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "DidUpdateRequest": {
        "type": "object",
        "description": "Request to update an existing DID document.",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DidDocument"
          },
          "logEntry": {
            "type": "object",
            "description": "New log entry with proof for the update"
          },
          "versionId": {
            "type": "string",
            "description": "Expected current version ID for optimistic locking"
          }
        },
        "required": ["document", "logEntry"]
      },
      "DidUpdateResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "versionId": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "VersionConflictResponse": {
        "type": "object",
        "description": "Returned when the server has a newer version of the DID document than the client.",
        "properties": {
          "error": { "type": "string" },
          "message": {
            "type": "string",
            "description": "Description of the conflict"
          },
          "serverVersionId": {
            "type": "string",
            "description": "Current version ID on the server"
          },
          "clientVersionId": {
            "type": "string",
            "description": "Version ID the client attempted to update from"
          },
          "serverDocument": {
            "$ref": "#/components/schemas/DidDocument",
            "description": "Current DID document on the server for client to merge"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Generic error response.",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "object" }
        }
      }
    }
  }
}